requirements.txt
python-fasthtml==0.4.4
uvicorn==0.30.1
seaborn
numpy
matplotlib
great_tables
